Move compilation to its own module
